requirements.txt
pandas
gspread
python-dotenv
openpyxl
airbyte
psycopg2-binary
sqlalchemy
sshtunnel
paramiko<4
